Clarify error message where dimension mappings are incorrect
The method validateDimensionMappings checks that the dimension mappings
between the given field and the signal field of the NXdata group being
built are valid. The number of dimension in the mapping must be
equal to the rank of the given field and each dimension in that mapping
must be a dimension of the signal field, i.e. between 0 and the
rank of the signal field (exclusively).

This change makes the exception message explain the problem more clearly,
including the names of both the signal field in the error message.



Change-Id: Id1708740e97ec4de0d4c83ea5aa4af9726d8a391