Removes incorrect jre.compilation.profile from bundles
jre.compilation.profile was set to Java 1.6 or Java 1.7 in several
bundles. This will fails any Tycho build as the
Bundle-RequiredExecutionEnvironment (BREE) in the MANIFEST.MF of the
affected bundles are already configured to use Java 8 and Tycho prefers
jre.compilation.profile over the BREE.

By removing it, we ensure that the BREE from the manifest is used bv
Tycho.

Change-Id: I649f6837733744d448b0ce8c858d9ea00e640dc5