[DAQ-2863] add method to get name of nexus device
By calling this method getName() it is already implemented by all
classes that implement INexusDevice. There is a slight danger that
a different name may should be returned, but this is unlikely. Calling
this method by a different name, e.g. getNexusDeviceName would mean that
this method would need to be added to all subclasses, probably just to
call whatever getName method that class already implemented.

Change-Id: I2a68136afd6f396a7ee649c5e1151535e8d19cca