[DAQ-2932] move NexusContext and associated classes from template plugin
Move NexusContext and implementations from org.eclipse.dawnsci.nexus.template
plugin to org.eclipse.dawnsci.nexus, so that they be used outside of
nexus templates. In particular, they are likely to be required for
implementing appenders that work for both classic and new (solstice)
scanning. This is because nexus writing for new scanning creates an
entire in-memory tree, and for classic scans writes each change to
disk.

Also adds a factory class, NexusContextFactory for creating NexusContexts.

Change-Id: I244478344a74c998f72bf2ba03512ca4be706e0e