DAQ-2838 Cleanup tests in org.eclipse.dawnsci.remotedataset.test
Some of these tests include specific workstation urls and expect
external processes to be running.

The deleted tests could possibly be improved/converted to be more like
unit tests but is likely not worth the effort.

A few tests were not run previously anyway since Buckminster used
the suite.

This change includes adding a start for HttpClient in RemoteDataset,
deleting Data Acquisition Development - DataServer as there is no
file/directory named 'dataserver' in the workspace and also deleting
the test suite so that tests are run in the usual way (now all present
tests should pass).

Change-Id: I1815548e2952be62badc50ba783e90491959b39e