[DAQ-2954] replace XPath concat function with || operator in XSLT files
XSLT 3.0 has now been released, which uses XPath 3.0. Among other new
features, this provides an infix string concatenation operator '||'
that we can use instead of the concat function to enhance readability.

Note: this change does not yet include the newly regenerated base
class and application definition classes as there are a number of issues
remaining to be fixed before this can be done without introducing errors

Change-Id: Ib8a655d033f845ebe1454f7d1447269247e15845