[DAQ-3325] do not create empty NXuser group as part of skeleton nexus tree
Since this nexus writing framework was created, it has created an
empty NXuser group with the name 'user' within the NXentry group to
every file. This group was not populated as we had not implemented a
way to write any details of the user (such as their user name or
their actual name) into the nexus tree.

Since the skeleton tree no longer contains an empty NXuser group, in
order to add a NXuser group to the tree (or multiple NXuser groups),
the NexusScanModel should contain one or more INexusDevices which
when asked for the their nexus objects return an NXuser group. Related
to this change, a new INexusDevice class that implements this behaviour
will added at the GDA layer that users the nexus writing API.

Change-Id: I84354009e5d172ac9e8a0450de0bc5e26d6dd64e