Add back org.eclipse.dawnsci.nexus.template.product
The product file was deleted as part of the build migration to Tycho.

Add the product back in a plain project which is imported into the
workspace via an Oomph component extension on the feature.

Replace EMF and ECF plugins with features since they are required
by org.eclipse.platform.

The product can be built (as part of the monolithic build) by running

 mvn -DskipTests -Pgda,nexus-template clean verify

Change-Id: Id9c7d9274ab943b1b5b0666dd01fc6e6656a5015