DAQ-3422 Fix length check in construction of RectangularROI.
Fix length check in construction of RectangularROI. The constructor
re-orients the ROI if a negative length is supplied (i.e. the bounds are
supplied in the wrong order) so that the object has positive length
values. It does so if the length of an axis is <= 0 but it is not
necessary to include 0 and it appears to cause problems.

Example case: new RectangularROI([0, 0], [0, 1]) looks like this: |, the
constructor would re-orient it to look like this: --

Change-Id: Idc413331b56f89986cb41044596c736bedae118a