[DAQ-3778] add getScanRole method to NexusObjectProvidera and remove NexusScanModel.multipleNexusDevice
The class NexusScanFileImpl which builds the nexus tree for a scan
treats NexusObjectProviders differently according to their ScanRole.
This affects how the fields within the nexus object are added to the
NXdata groups.

The NexusObjectProviders are extracted from the
INexusDevices in the NexusScanModel, which are already grouped by
ScanRole. However, an INexusDevice may control multiple real devices
with different roles in the scan. The only current example of this is
a malcolm device (IMalcolmDevice).

To handle this, the NexusScanModel previously had a special field for
this called multipleNexusDevice, which it had special handling for - it
asked the NexusObjectProviders for the type of nexus object they created
and added the providers to the list of providers for the appropriate
ScanRole. However, this was at best a guess.

The new solution is to add a method NexusObjectProvider.getScanRole(),
this property can be used to specify the role of a NexusObjectProvider
in the scan, overriding the ScanRole by which the INexusDevice was
grouped with as part of the NexusScanModel

Change-Id: I80264302d9581d25131990c80ed178e9829b74e1