[DAQ-3786] fix starting up services for NexusValidationApplication
For some reason org.eclipse.dawnsci.nexus.ServiceHolder cannot be
created when running the nexus validation application. The problem is
with the dependency on INexusFileFactory - commenting out that line
works, but we then need to get that service from the OSGi BundleContext.
The ServiceHolder can be created as normal when starting up the
GDA client and server.

Also, added try/catch to show any exception that occurred when trying
to validate a nexus file.

Change-Id: Id15d42eaa788c7cfee01f2dd7d45e87a87aa3c42